+++ components/form/types/date.type.tsx
@@ -2,12 +2,14 @@ import { Form } from 'antd'
 import dayjs, { Dayjs } from 'dayjs'
 import moment, { Moment } from 'moment'
 import React, { useEffect, useState } from 'react'
+import { useTranslation } from 'react-i18next'
 import { StyledDateInput } from '../../styled/date.input'
 import { FieldTypeProps } from './type.props'
 
 export const DateType: React.FC<FieldTypeProps> = ({ field, design }) => {
   const [min, setMin] = useState<Dayjs>()
   const [max, setMax] = useState<Dayjs>()
+  const { t } = useTranslation()
 
   useEffect(() => {
     field.options.forEach((option) => {
@@ -24,7 +26,7 @@ export const DateType: React.FC<FieldTypeProps> = ({ field, design }) => {
     <div>
       <Form.Item
         name={[field.id, 'value']}
-        rules={[{ required: field.required, message: 'Please provide Information' }]}
+        rules={[{ required: field.required, message: t('validation:valueRequired') }]}
         getValueFromEvent={(e: Moment) => e.format('YYYY-MM-DD')}
         getValueProps={(e: string) => ({ value: e ? moment(e) : undefined })}
         initialValue={field.value ? moment(field.value) : undefined}