+++ components/form/types/email.type.tsx
@@ -1,16 +1,19 @@
 import { Form } from 'antd'
 import React from 'react'
+import { useTranslation } from 'react-i18next'
 import { StyledInput } from '../../styled/input'
 import { FieldTypeProps } from './type.props'
 
 export const EmailType: React.FC<FieldTypeProps> = ({ field, design }) => {
+  const { t } = useTranslation()
+
   return (
     <div>
       <Form.Item
         name={[field.id, 'value']}
         rules={[
-          { required: field.required, message: 'Please provide Information' },
-          { type: 'email', message: 'Must be a valid email' },
+          { required: field.required, message: t('validation:valueRequired') },
+          { type: 'email', message: t('validation:invalidEmail') },
         ]}
         initialValue={field.value}
       >