+++ components/form/types/link.type.tsx
@@ -1,16 +1,19 @@
 import { Form } from 'antd'
 import React from 'react'
+import { useTranslation } from 'react-i18next'
 import { StyledInput } from '../../styled/input'
 import { FieldTypeProps } from './type.props'
 
 export const LinkType: React.FC<FieldTypeProps> = ({ field, design }) => {
+  const { t } = useTranslation()
+
   return (
     <div>
       <Form.Item
         name={[field.id, 'value']}
         rules={[
-          { required: field.required, message: 'Please provide Information' },
-          { type: 'url', message: 'Must be a valid URL' },
+          { required: field.required, message: t('validation:valueRequired') },
+          { type: 'url', message: t('validation:invalidUrl') },
         ]}
         initialValue={field.value}
       >