+++ components/form/types/text.type.tsx
@@ -1,16 +1,18 @@
 import { Form } from 'antd'
 import React from 'react'
+import { useTranslation } from 'react-i18next'
 import { StyledInput } from '../../styled/input'
 import { FieldTypeProps } from './type.props'
 
 export const TextType: React.FC<FieldTypeProps> = ({ field, design }) => {
+  const { t } = useTranslation()
   // TODO focus when becomes visible
 
   return (
     <div>
       <Form.Item
         name={[field.id, 'value']}
-        rules={[{ required: field.required, message: 'Please provide Information' }]}
+        rules={[{ required: field.required, message: t('validation:valueRequired') }]}
         initialValue={field.value}
       >
         <StyledInput design={design} allowClear size={'large'} />